Update MMM-SL.js - New SL API
Updated the code to interface with the new SL API for Departures https://transport.integration.sl.se/v1/sites

Also removed the MMM-Scheduler functionality as it is not required any longer. The SL API allows more requests.

Also modified the departures to use the name from stop_point name instead of stop_area name. In addition when displaying multple stops the names are always sorted in alphabetical order to keep the display consistent.